handle only framework parameter
