+++ cmd/make-js-component.mjs
@@ -1,5 +1,6 @@
 #! /usr/bin/env node
-import wizard from "../utils/wizard/wizard.mjs";
+//const wizard = require('../utils/wizard.mjs');
+import wizard from "../utils/wizard.mjs";
 import createComponent from "../src/utils/utils.mjs";
 wizard().then((answers) => {
     const { componentName, framework, template, folder } = answers;
