--- types/index.mjs
@@ -1,1 +1,0 @@
-export {};
