--- types/index.mts
@@ -1,11 +1,0 @@
-
-type Answers = {
-    componentName: string,
-    framework: string,
-    template: string,
-    folder: string
-};
-
-export {
-    Answers
-}
