--- utils/wizard/frameworks/angular.wizard.mjs
@@ -1,9 +1,0 @@
-const angularWizard = (componentName, folder) => {
-    return {
-        componentName,
-        folder,
-        framework: "angular",
-        template: "component.component.js",
-    };
-};
-export default angularWizard;
